# api-getway/api/docs/swagger.yaml
definitions:
  models.AdminPost:
    properties:
      category:
        type: string
      content:
        type: string
      id:
        type: string
      image-url:
        type: string
      owner_id:
        type: string
      title:
        type: string
    type: object
  models.Comment:
    properties:
      description:
        type: string
      id:
        type: string
      postId:
        type: string
    type: object
  models.Error:
    properties:
      message:
        type: string
    type: object
...
  models.StandardErrorModel:
    properties:
      error:
        $ref: '#/definitions/models.Error'
    type: object
  models.User:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
host: localhost:1212
info:
  contact: {}
  title: welcome to
  version: "1.7"
paths:
...
  /v1/admin/create/post/:
    post:
      consumes:
      - application/json
      description: Api for creating a new post
      parameters:
      - description: Create post
        in: body
        name: Post
        required: true
        schema:
          $ref: '#/definitions/models.AdminPost'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AdminPost'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      security:
      - ApiKeyAuth: []
      summary: AdminCreatePost
      tags:
      - admin
  /v1/admin/delete/comment:
    delete:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Comment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      security:
      - ApiKeyAuth: []
      summary: AdminDeleteComment
      tags:
      - admin
  /v1/admin/delete/user:
    delete:
      consumes:
      - application/json
      description: Api for deleting user by id
      parameters:
      - description: ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      security:
      - ApiKeyAuth: []
      summary: AdminDeleteUser
      tags:
      - admin
  /v1/admin/update/post/:
    put:
      consumes:
      - application/json
      description: Api for creating a new post
      parameters:
      - description: update post
        in: body
        name: Post
        required: true
        schema:
          $ref: '#/definitions/models.AdminPost'
      - description: id
        in: query
        name: post_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AdminPost'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      security:
      - ApiKeyAuth: []
      summary: AdminUpdatePost
      tags:
      - admin
  /v1/admin/update/user/:
    put:
      consumes:
      - application/json
      description: Api for updating user
      parameters:
      - description: createUserModel
        in: body
        name: User
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      security:
      - ApiKeyAuth: []
      summary: AdminUpdateUser
      tags:
      - admin
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"